+++ CHANGELOG.md
@@ -183,6 +183,8 @@ const SELECTORS = {
 - Changed `createItem()` to return data in the model-form as `FormService` will convert it to state
 - Remove `clipErrors` as the existence of `validators` will clip them going forward
 - `validators` can be applied to a selector as long as none of their ancestor selectors apply `validators`
+- Added `validateRaw` to unformat value
+- Added `validateManually`
 - Throw an error if `validators` is defined on selectors with parent selectors that also define `validators`
 - Throw an error if calling `apply()` on a key where `pristine` is type `object`
 - Validation workflow with `apply()`:
@@ -193,4 +195,3 @@ const SELECTORS = {
 ## v2.1 (Pre-Release)
 
 - `createItem()` will pass parameters like so: `createItem(path, index, state, service, opts)`
-- Add `manualValidation` flag